Fix #138.
SRPMImportCommand failed with a bad error message when an SRPM is
imported into twice into a project or when an SRPM with the exact name
of the imported SRPM already existed in the project. This fix improves
the error message and provides hints as to what might have gone wrong.