Fix #134.
A NPE was thrown due to attempting to dereference the MockBuildResult
instance in the job done listener (the MockBuildCommand throws an
exception preventing the call() to finish successfully).

We now create a custom status in the relevant jobs and check for this
status in the job done listener to simply return in the case of mock not
being installed or the user not being a member of the mock group.