Fix #123.
Koji SRPM based scratch builds now don't prompt for the SRPM if the user
right-clicks on an SRPM. If the event source was not an SRPM the handler
behaves in the same way it did prior to this fix. I.e. give the user a
list of SRPMs in the project (if any) or provide a file dialog as a last
resort.